sanctum auth Bearer Token & Protected routes & Gates

+++ event-management/app/Http/Controllers/Api/AttendeeController.php
@@ -8,6 +8,7 @@ use App\Http\Traits\CanLoadRelationships;
 use App\Models\Attendee;
 use Illuminate\Http\Request;
 use App\Models\Event;
+use Illuminate\Support\Facades\Gate;
 
 class AttendeeController extends Controller
 {
@@ -17,6 +18,7 @@ class AttendeeController extends Controller
    
     public function index(Event $event)
     {
+
         $attendees = $this->loadRelationships(
             $event->attendees()->latest()
         );
@@ -52,8 +54,9 @@ class AttendeeController extends Controller
     }
 
  
-    public function destroy(string $event, Attendee $attendee)
+    public function destroy(Event $event, Attendee $attendee)
     {
+        Gate::authorize('delete-attendee', [$event, $attendee]);
         $attendee->delete();
 
         return response(status: 204);
